changes based on comments #3764
